Redo long query handling
Simply truncating the query can yield false-positives so
when the query exceeds max-len, do an additional pass
on the results with the helm matcher to weed them out.

This way, the preferred matches have a good chance
of appearing in the final list instead of disappearing
when the query length exceeds max-len.

Normally, a long-ish query would be sufficient to produce
a number of matches below LIMIT and so this probably doesn't
make a huge difference, but in some cases the results should
be better.