Fix breakage caused by helm b1ec2d1a
Helm eliminated the external hash table for deduping matches
and cleaned up the code for building up the candidate list.
Still incompatible with fuzzier calling helm's logic after
already having selected some matches, so we use the new function
from helm but modify it to use a global hash table (again)
which we reintroduce, this time under the fuzzier namepsace.

Closes #12